[MPT-181] Moving times to dedicated directory
