+++ packages/GIS-ly/ui/src/model/GIS/Time/TimeRange.ts
